release time test

+++ src/forest_fire_detection_system/src/app/InMotionDropping/InMotionDropping_Flight_control_node.cpp
@@ -590,8 +590,9 @@ void velocityAndYawRateCtrl(const JoystickCommand &offsetDesired, uint32_t timeM
 {
   double originTime  = 0;
   double currentTime = 0;
-  uint64_t elapsedTimeInMs = 0;
-  
+  // uint64_t elapsedTimeInMs = 0;
+   float elapsedTimeInMs = 0;
+
   SetJoystickMode joystickMode;
   JoystickAction joystickAction;
 
@@ -615,7 +616,7 @@ void velocityAndYawRateCtrl(const JoystickCommand &offsetDesired, uint32_t timeM
   {
     currentTime = ros::Time::now().toSec();
     elapsedTimeInMs = (currentTime - originTime) * 1000;
-      ROS_INFO("timeinMs [%d]",elapsedTimeInMs);
+      ROS_INFO("timeinMs [%f]",elapsedTimeInMs);
       joystick_action_client.call(joystickAction);
     if(elapsedTimeInMs>1000){
         ROS_INFO("release valve");
